+++ backend/src/resource/asset/index.html
@@ -3,10 +3,6 @@
         <title>70年代、80年代、90年代，怀旧、复古、回忆</title>
         <link rel="stylesheet" href="/asset/bootstrap.min-6c05f7396999b9df.css">
         <link rel="stylesheet" href="/asset/bootstrap-icons-23ab5ed540459d1e.css">
-        
-        
-        
-        
-    <script type="module">import init from '/asset/index-2cb1cb1112c4b0f5.js';init('/asset/index-2cb1cb1112c4b0f5_bg.wasm');</script></head>
+    <script type="module">import init from '/asset/index-a844d9f5ca06acdf.js';init('/asset/index-a844d9f5ca06acdf_bg.wasm');</script></head>
     <body>
 </body></html>
