+++ backend/src/service/asset_list.rs
@@ -4,8 +4,8 @@
 ("codemirror.min.css", include_bytes!(r#"..\resource\asset/codemirror.min.css.gz"#), "text/css"),
 ("fonts/bootstrap-icons.woff", include_bytes!(r#"..\resource\asset/fonts\bootstrap-icons.woff.gz"#), ""),
 ("fonts/bootstrap-icons.woff2", include_bytes!(r#"..\resource\asset/fonts\bootstrap-icons.woff2.gz"#), ""),
-("index-2cb1cb1112c4b0f5.js", include_bytes!(r#"..\resource\asset/index-2cb1cb1112c4b0f5.js.gz"#), "text/javascript"),
-("index-2cb1cb1112c4b0f5_bg.wasm", include_bytes!(r#"..\resource\asset/index-2cb1cb1112c4b0f5_bg.wasm.gz"#), "application/wasm"),
+("index-a844d9f5ca06acdf.js", include_bytes!(r#"..\resource\asset/index-a844d9f5ca06acdf.js.gz"#), "text/javascript"),
+("index-a844d9f5ca06acdf_bg.wasm", include_bytes!(r#"..\resource\asset/index-a844d9f5ca06acdf_bg.wasm.gz"#), "application/wasm"),
 ("index.html", include_bytes!(r#"..\resource\asset/index.html.gz"#), "text/html; charset=utf-8"),
 ("snippets/blog-frontend-3607fd663da6040e/asset/editor.js", include_bytes!(r#"..\resource\asset/snippets\blog-frontend-3607fd663da6040e\asset\editor.js.gz"#), "text/javascript"),
 ("toastui-editor-all.min.js", include_bytes!(r#"..\resource\asset/toastui-editor-all.min.js.gz"#), "text/javascript"),
